Separated IDP solvers from remhos.cpp.
