+++ remhos.cpp
@@ -42,6 +42,7 @@
 #include "remhos_mono.hpp"
 #include "remhos_tools.hpp"
 #include "remhos_sync.hpp"
+#include "remhos_solvers.hpp"
 
 using namespace std;
 using namespace mfem;
@@ -77,16 +78,6 @@ double inflow_function(const Vector &x);
 
 // Mesh bounding box
 Vector bb_min, bb_max;
-
-class RK2IDPSolver : public ODESolver
-{
-   Vector dx12, dx;
-
-public:
-   RK2IDPSolver();
-   void Init(TimeDependentOperator &f) override;
-   void Step(Vector &x, double &t, double &dt) override;
-};
 
 class AdvectionOperator : public TimeDependentOperator
 {
@@ -1872,29 +1863,4 @@ double inflow_function(const Vector &x)
       return 0.25*(1.+tanh((r+c-a)/b))*(1.-tanh((r-c-a)/b));
    }
    else { return 0.0; }
-}
-
-RK2IDPSolver::RK2IDPSolver()
-{
-}
-
-void RK2IDPSolver::Init(TimeDependentOperator &f)
-{
-   ODESolver::Init(f);
-   dx12.SetSize(f.Height());
-   dx.SetSize(f.Height());
-}
-
-void RK2IDPSolver::Step(Vector &x, double &t, double &dt)
-{
-   f->SetTime(t);
-   f->Mult(x, dx12);
-
-   x.Add(dt/2., dx12);
-   f->SetTime(t+dt/2.);
-   f->Mult(x, dx);
-
-   add(2., dx, -1., dx12, dx);
-
-   x.Add(dt/2., dx);
 }
